% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downlaod_utils.R
\name{baixar_captchas}
\alias{baixar_captchas}
\title{Baixar captchas}
\usage{
baixar_captchas(n, url = url(), folder)
}
\arguments{
\item{n}{quantidade de captchas}

\item{url}{do capctha}

\item{folder}{pasta para salvar os capcthas baixados}
}
\description{
Baixar captchas
}

