% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{saj_baixar}
\alias{saj_baixar}
\title{Baixar captcha}
\usage{
saj_baixar(files, tipo = "both", tj = "sp")
}
\arguments{
\item{files}{arquivos para salvar os captchas. Deve ter tamanho 1 se tipo for audio ou image e deve ter tamanho 2 se tipo for both.}

\item{tipo}{"audio" para audio e "image" para imagem, 'both' para ambos. Default "both".}

\item{tj}{TJ do captcha}
}
\description{
Baixar um captcha dependendo do TJ
}

