% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{saj_baixar_captchas}
\alias{saj_baixar_captchas}
\title{Baixar e classificar captchas de imagem.}
\usage{
saj_baixar_captchas(N, baixar_audio = FALSE, tj = "sp", folder,
  folder_audio = NULL)
}
\arguments{
\item{N}{quantidade de captchas}

\item{baixar_audio}{se \code{TRUE} baixa o áudio além da imagem (necessário especificar \code{folder_audio}).}

\item{tj}{Tribunal de origem do captcha (por enquanto só "sp" e "sc").}

\item{folder}{pasta em que os captchas serão salvos.}

\item{folder_audio}{pasta para salvar os áudios.}
}
\description{
Laço criado para baixar e classificar vários captchas de imagem e salvar numa pasta.
}

