% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{saj_classificar_auto}
\alias{saj_classificar_auto}
\title{Classifica automaticamente N captchas}
\usage{
saj_classificar_auto(N, folder_img = getwd(), folder_aud = folder_img,
  verbose = TRUE, intervalo = 50L)
}
\arguments{
\item{N}{número de captchas a serem baixados e classificados automaticamente.}

\item{folder_img}{pasta onde os arquivos de imagem serão salvos. Default para pasta atual.}

\item{folder_aud}{pasta onde os arquivos de áudio serão salvos. Default para \code{folder_img}.}

\item{verbose}{imprimir mensagens de acompanhamento do número de captchas baixados.}

\item{intervalo}{número de captchas baixados para gerar mensagens caso \code{verbose == TRUE}.}
}
\description{
Utiliza a função \code{\link[captchaSajAudio]{decifrar}} que acerta praticamente 100\% dos áudios. A função baixa tanto os captchas de áudio quanto de imagem.
}

